refactor: menu listings a bit
